+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2024-12-16
+
+### Added
+
+- `[RequiredField]` attribute that displays an error indicator in the inspector and scene hierarchy if the attached field is left empty
+    - Currently functions for empty strings and missing Object references
+
 ## [0.3.1] - 2024-12-01
 
 ### Added